[fix] 코드리뷰 반영
